# contract/openapi.yaml
openapi: 3.1.0
info:
  title: PaytmFlow API
  version: "1.0.0"
  description: >
    Canonical API contract for PaytmFlow (spec v4.2).
    FROZEN at T+3. Changes require the contract-change protocol in SHARED_CONTRACT.md §9.
    Dev2 owns this file. Dev1 generates TypeScript types from it and must never hand-edit them.
servers:
  - url: http://localhost:8000
    description: Local
  - url: /
    description: Deployed (same origin via proxy)
...
paths:
...
  /api/v1/chat:
    post:
      tags: [journeys]
      operationId: generalChat
      description: >
        Advisory-only Q&A with NO journey context - used before any application
        exists (e.g. the goal-creation form). Never claims to know specifics about
        the caller's own application, since there isn't one yet.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChatRequest" }
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatResponse" }
  /api/v1/journeys/{journey_id}/chat:
    post:
      tags: [journeys]
      operationId: chatWithAssistant
      description: >
        Advisory-only Q&A about the caller's own journey. Grounded in the same
        server-computed journey_state/recommendation data the status/recommendation
        screens already render. Never writes state.
      parameters: [{ $ref: "#/components/parameters/JourneyIdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChatRequest" }
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/journeys/{journey_id}/chat/voice:
    post:
      tags: [journeys]
      operationId: chatWithAssistantVoice
      description: >
        Voice input for the same advisory-only assistant as chatWithAssistant -
        the audio is transcribed (Sarvam Saaras) and the transcript is fed into
        the SAME grounded chat() call a typed message would use. Voice is an
        input modality, not an authorization mechanism: it can never write
        journey state.
      parameters: [{ $ref: "#/components/parameters/JourneyIdPath" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                language_code: { type: string, default: unknown }
      responses:
        "200":
          description: Transcript and assistant reply
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VoiceChatResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "504":
          description: Transcription timed out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pf_session
    demoSecret:
      type: apiKey
      in: header
      name: X-Demo-Secret
  parameters:
    JourneyIdPath:
      name: journey_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: Not found, or not owned by this session (deliberately indistinguishable)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationError:
      description: VALIDATION_ERROR or INVALID_JOURNEY_TYPE
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - INVALID_JOURNEY_TYPE
        - VALIDATION_ERROR
        - ACTION_INVALID
        - ACTION_STALE
        - NEEDS_REVIEW
        - DEAD_END
        - AI_TIMEOUT
        - AI_MALFORMED_OUTPUT
        - EVIDENCE_CONFLICT
        - NOT_FOUND
        - FORBIDDEN
        - REVIEW_CASE_STALE
        - REVIEW_CASE_LOCKED
        - REVIEW_CASE_INVALID_TRANSITION
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Human-readable and safe to render directly to the user.
            details:
              type: object
              additionalProperties: true
            current_snapshot_id:
              type: string
              format: uuid
              description: Present on ACTION_STALE so the client can resync without a second call.
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1, maxLength: 500 }
    ChatResponse:
      type: object
      required: [reply]
      properties:
        reply: { type: string }
    VoiceChatResponse:
      type: object
      required: [reply, transcript]
      properties:
        reply: { type: string }
        transcript:
          type: string
          description: >
            The Sarvam Saaras transcription the assistant actually answered
            from, shown back to the customer for confirmation.
        language_code: { type: string, nullable: true }
...
security:
  - sessionCookie: []